% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MahalanobisDis.R
\name{CalculateMD}
\alias{CalculateMD}
\title{Calculate Mahalanobis distance}
\usage{
CalculateMD(clr.norm, core.drop.assign, kmed.cl, cluster.assign)
}
\arguments{
\item{clr.norm}{The local CLR normalized data comes from the output of the 
\link{LocalCLRNorm} function.}

\item{core.drop.assign}{Classification of core and non-core cells from the 
output of the \link{DefineNonCore} function.}

\item{kmed.cl}{K-Medoids clustering results from the output of the 
\link{KmedCluster} function.}

\item{cluster.assign}{The cluster assignments and corresponding hashtags from 
the output of \link{LabelClusterHTO}.}
}
\value{
A matrix of Mahalanobis distances, where rows correspond to hashtags 
and columns correspond to cells.
}
\description{
Calculate the Mahalanobis distance of each cell to its cluster centroid.
}
\details{
The covariance matrix \eqn{\Gamma} between pairs of hashtags (\eqn{i,i'}) is
\deqn{
\Gamma_{i,i'} = \frac{\sum_l \sum_{c \in l} (h_{i,c} - M_i^{(l)}) (h_{i',c} - M_{i'}^{(l)})}{\sum_l n_l},
}
Here, \eqn{n_l} denotes the number of cells in cluster \eqn{l}, and \eqn{h_{i,c}} 
denotes the local CLR value of hashtag \eqn{i} in cell \eqn{c}.

Since \eqn{\Gamma} is singular and not invertible, we use the Moore-Penrose 
generalized inverse, denoted \eqn{\Gamma^+}, to compute a pseudo-inverse of 
\eqn{\Gamma}.

The Mahalanobis distance \eqn{MD_{c,l}} for each cell \eqn{c} to the centroid 
\eqn{M^{(l)}} of the K-medoids cluster \eqn{l} is defined as
\deqn{
MD_{c,l} = (h_c - M^{(l)})^T \Gamma^+ (h_c - M^{(l)}), \quad
h_c = (h_{1,c}, \dots, h_{N,c})^T.
}
}
\examples{
\donttest{
hto.md.mat <- CalculateMD(hto.cl.mtx, hto.noncore, hto.kmed.cl, hto.cluster.assign)
dim(hto.md.mat)
hto.md.mat[,1:10]
}

}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S-PLUS. 
Fourth Edition. Springer.

Mahalanobis, P. C. (2018). On the generalized distance in statistics. 
SankhyÄ: The Indian Journal of Statistics, Series A (2008-), 80, S1-S7.
}
