% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{CheckAssign2DPlot}
\alias{CheckAssign2DPlot}
\title{Check suspicious demultiplexed cells}
\usage{
CheckAssign2DPlot(
  hash.count,
  gex.count = NULL,
  md.mat,
  cmddemux.assign,
  check_barcodes = NULL,
  check_barcodes_col = "blue",
  check_barcode_size = 2
)
}
\arguments{
\item{hash.count}{The raw hashing count matrix has rows representing hashtags 
and columns representing cells.}

\item{gex.count}{A gene expression count matrix with rows representing genes 
and columns representing cells. Only required when gene expression data
is available. Default is NULL.}

\item{md.mat}{Matrix of Mahalanobis distance from the output of the 
\link{CalculateMD} function.}

\item{cmddemux.assign}{The demultiplexing results are returned by the 
\link{CMDdemuxClass} function.}

\item{check_barcodes}{A character vector of cell barcodes to be examined. 
Default: NULL.}

\item{check_barcodes_col}{The color used to highlight the specified cells 
in the plot. Default: "blue".}

\item{check_barcode_size}{The size of the points representing the specified 
cells in the plot. Default: 2.}
}
\value{
A combined layout of pairwise information from hashtag library sizes, 
mRNA library sizes, the 2nd MD, and the ratio between the 1st MD and the 2nd 
MD. Cells are color-coded by classification (singlets, negatives, doublets), 
with selected barcodes highlighted.
}
\description{
Use pairwise information from hashtag library sizes, mRNA library sizes, 
the second Mahalanobis distance (2nd MD), and the ratio between the minimum 
Mahalanobis distance (1st MD) and the 2nd MD to assess whether droplets are 
correctly classified.
}
\examples{
\donttest{
hto.check.barcodes <- colnames(hto.mtx)[1:30]
CheckAssign2DPlot(hto.mtx, hto.md.mat, hto.cmddemux.assign, gex.count = hto.gex.count, check_barcodes = hto.check.barcodes)
}

}
