% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{CheckAssign3DPlot}
\alias{CheckAssign3DPlot}
\title{Check suspicious demultiplexed cells in a 3D plot}
\usage{
CheckAssign3DPlot(
  hash.count,
  gex.count,
  md.mat,
  cmddemux.assign,
  check_barcodes = NULL,
  check_barcodes_col = "blue",
  point_size = 2,
  check_barcode_size = 5
)
}
\arguments{
\item{hash.count}{The raw hashing count matrix has rows representing hashtags 
and columns representing cells.}

\item{gex.count}{A gene expression count matrix with rows representing genes 
and columns representing cells. Only required when gene expression data
is available.}

\item{md.mat}{Matrix of Mahalanobis distance from the output of the 
\link{CalculateMD} function.}

\item{cmddemux.assign}{The demultiplexing results are returned by the 
\link{CMDdemuxClass} function.}

\item{check_barcodes}{A character vector of cell barcodes to be examined. 
Default: NULL.}

\item{check_barcodes_col}{The color used to highlight the specified cells 
in the plot. Default: "blue".}

\item{point_size}{The size of points in the plot. Default: 2.}

\item{check_barcode_size}{The size of the points representing the specified 
cells in the plot. Default: 5.}
}
\value{
A 3D plot with library sizes, mRNA library sizes, and the ratio of the minimum to the second 
Mahalanobis distance. Cells are color-coded by classification (singlets, 
negatives, doublets), with selected barcodes highlighted.
}
\description{
Assess the accuracy of droplet classification by visualizing hashtag library 
sizes, mRNA library sizes, and the ratio of the minimum to the second 
Mahalanobis distance.
}
\examples{
\donttest{
hto.check.barcodes <- colnames(hto.mtx)[1:30]
CheckAssign3DPlot(hto.mtx, hto.gex.count, hto.md.mat, hto.cmddemux.assign, hto.check.barcodes)
}

}
