% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{CheckCLRBoxPlot}
\alias{CheckCLRBoxPlot}
\title{Check the number of clusters with a boxplot}
\usage{
CheckCLRBoxPlot(clr.norm, kmed.cl)
}
\arguments{
\item{clr.norm}{The local CLR normalized data comes from the output of the 
\link{LocalCLRNorm} function.}

\item{kmed.cl}{K-Medoids clustering results from the output of \link{KmedCluster} 
function.}
}
\value{
A combined layout of boxplots, where each panel represents the 
expression of a specific hashtag across clusters.
}
\description{
Use a boxplot to assess whether the chosen number of clusters is appropriate. 
Ideally, one box should be significantly higher than the others in each panel.
}
\examples{
\donttest{
CheckCLRBoxPlot(hto.clr.mtx, hto.kmed.cl)
}

}
