% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{CLRPlot}
\alias{CLRPlot}
\title{Scatter plot of pairwise local CLR values}
\usage{
CLRPlot(
  clr.norm,
  kmed.cl,
  core.drop.assign,
  return_cluster_only = TRUE,
  seed = 10000
)
}
\arguments{
\item{clr.norm}{The local CLR normalized data comes from the output of the 
\link{LocalCLRNorm} function.}

\item{kmed.cl}{K-Medoids clustering results from the output of \link{KmedCluster} 
function.}

\item{core.drop.assign}{Classification of core and non-core cells from the 
output of the \link{DefineNonCore} function.}

\item{return_cluster_only}{Logical. If TRUE, one plot is randomly selected per 
cluster. If FALSE, plots of all pairwise local CLR values (h) are returned. 
Default: TRUE.}

\item{seed}{Set a random seed. Default: 10000.}
}
\value{
A combined layout of scatter plots of pairwise local CLR values (h), 
where core cells are colored by cluster and non-core cells are shown in grey. 
Cluster medoids are indicated by cross markers.
}
\description{
Scatter plots of pairwise local CLR values, where core cells are colored by 
cluster, non-core cells are shown in grey, and cluster medoids are indicated.
}
\examples{
\donttest{
CLRPlot(hto.clr.mtx, hto.kmed.cl, hto.noncore)
}

}
