% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{DefineNonCore}
\alias{DefineNonCore}
\title{Define non-core cells}
\usage{
DefineNonCore(
  cl.dist,
  kmed.cl,
  eu_cut_q = 0.9,
  eu_cut = NULL,
  optional = FALSE,
  clr.norm = NULL
)
}
\arguments{
\item{cl.dist}{The Euclidean distance for each cell to its cluster centroid 
from the output of \link{EuclideanClusterDist} function.}

\item{kmed.cl}{K-Medoids clustering results from the output of 
\link{KmedCluster} function.}

\item{eu_cut_q}{A vector of cut-offs for defining non-core cells using the 
quantile of Euclidean distances within each cluster. Must be the same 
length as the number of clusters. This parameter can be explored by 
visualizing the distribution of Euclidean distances using the 
\link{EuclideanDistPlot} function. Default: 0.9.}

\item{eu_cut}{A vector of cut-offs for defining non-core cells using the raw 
Euclidean distance within each cluster. Must be the same length as the 
number of clusters. Use either \code{eu_cut} or \code{eu_cut_q} to 
determine the cut-offs of non-core cells. Default: NULL.}

\item{optional}{Logical; whether optional clustering has been performed to 
obtain more clusters than the number of hashtags. This parameter should 
be the same as the \code{optional} parameter in the \link{KmedCluster} 
function. Default: FALSE.}

\item{clr.norm}{Local CLR-normalized data from the output of \link{LocalCLRNorm} 
function. This parameter only needs to be set when \code{optional = TRUE}. 
Default: NULL.}
}
\value{
A vector of cell labels. Non-core cells are labeled as "non-core", and 
core cells are labeled with their corresponding cluster.
}
\description{
Non-core cells are defined based on their Euclidean distances to cluster centroids.
}
\details{
Core cells are those with smaller Euclidean distances to their cluster centroids, 
whereas non-core cells have larger Euclidean distances. The cut-off for defining 
non-core cells can be based either on the quantile of Euclidean distances or 
on a raw Euclidean distance value. Cells with Euclidean distances larger than 
the cut-off in each cluster are defined as non-core cells. 

When optional clustering is performed, cells in additional clusters that do not 
correspond to any hashtag are also defined as non-core cells. Each hashtag should 
have its maximum CLR expression in one cluster. Extra clusters that do not show 
maximum CLR expression for any hashtag are therefore classified as non-core, 
and all cells within them are considered non-core cells.
}
\examples{
\donttest{
hto.kmed.cl1 <- KmedCluster(hto.clr.mtx)
hto.noncore <- DefineNonCore(hto.cl.list, hto.kmed.cl1, c(0.9, 0.75, 0.94))
hto.kmed.cl2 <- KmedCluster(hto.clr.mtx, optional = TRUE, extra_cluster = 1)
# Define non-core cells in the data using optional clustering
hto.noncore <- DefineNonCore(hto.cl.list, hto.kmed.cl2, c(0.9, 0.95, 0.83, 0.65), optional = TRUE, clr.norm = hto.clr.norm)
head(hto.noncore)
}

}
