% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMedoidsCluster.R
\name{EuclideanClusterDist}
\alias{EuclideanClusterDist}
\title{Euclidean distance to the cluster centroid}
\usage{
EuclideanClusterDist(clr.norm, kmed.cl)
}
\arguments{
\item{clr.norm}{The local CLR normalized data comes from the output of 
\link{LocalCLRNorm} function.}

\item{kmed.cl}{K-Medoids clustering information from the output of
\link{KmedCluster} function.}
}
\value{
A list of Euclidean distances. The length of the list equals the number 
of clusters, and each element corresponds to the Euclidean distances of the 
cells in that cluster.
}
\description{
Calculate the Euclidean distance from each cell to its assigned cluster centroid.
}
\details{
The Euclidean distance \eqn{d} between each cell \eqn{c} to its cluster \eqn{l}
centroid \eqn{M^{(l)}} is:
\eqn{d = \| h_c - M^{(l)} \|_2, \quad c \in l}
}
\examples{
\donttest{
hto.cl.dist <- KmedCluster(hto.clr.mtx, hto.kmed.cl)
length(hto.cl.dist)
head(hto.cl.dist[[1]])
}   

}
