% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{EuclideanDistPlot}
\alias{EuclideanDistPlot}
\title{Density plot of Euclidean distances}
\usage{
EuclideanDistPlot(cl.dist, eu_cut_q = 0.9, eu_cut = NULL)
}
\arguments{
\item{cl.dist}{The Euclidean distance for each cell to its cluster centroid 
from the output of \link{EuclideanClusterDist} function.}

\item{eu_cut_q}{A vector of cut-offs for defining non-core cells using the 
quantile of Euclidean distances within each cluster. Must be the same 
length as the number of clusters. The red line in the plot represents 
the defined cut-off. Default: 0.9.}

\item{eu_cut}{A vector of cut-offs for defining non-core cells based on the 
within-cluster Euclidean distances. Must be the same length as the number 
of clusters. This only needs to be specified when the \code{eu_cut_q} 
parameter cannot provide suitable cut-offs. The red line in the plot 
represents the defined cut-off. Default: NULL.}
}
\value{
A combined layout of density plots, where each plot shows the distribution 
of Euclidean distances for a cluster.
}
\description{
Draw a density plot of the Euclidean distance distribution for cells in each 
cluster.
}
\examples{
\donttest{
EuclideanDistPlot(hto.cl.dist, c(0.9, 0.9, 0.75, 0.92, 0.95, 0.94, 0.95, 0.92))
}

}
