% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMedoidsCluster.R
\name{ExamineClusterNumber}
\alias{ExamineClusterNumber}
\title{Examine the number of clusters}
\usage{
ExamineClusterNumber(clr.norm, kmed.cl1, kmed.cl2, fc_cut = 1)
}
\arguments{
\item{clr.norm}{The local CLR normalized data comes from the output of 
\link{LocalCLRNorm} function.}

\item{kmed.cl1}{K-Medoids clustering results with fewer clusters (smaller k) 
from the output of the \link{KmedCluster} function.}

\item{kmed.cl2}{K-Medoids clustering results with more clusters (larger k) 
from the output of the \link{KmedCluster} function.}

\item{fc_cut}{Cut-off for determining the number of clusters. It is defined as 
the log fold change of the CLR expression of a specific hashtag between 
the top clusters for two k values. Default: 1.}
}
\value{
A message is printed indicating the largest CLR difference, its value, 
and the number of clusters selected.
}
\description{
Examine the number of clusters that should be used in the data.
}
\details{
For each hashtag \eqn{i}, the median local CLR value \eqn{h} across all clusters 
\eqn{l} is defined as
\deqn{med_{i,l} = median_{c \in l}\{ h_{i,c} \}}

The top cluster with the largest median CLR is defined as
\deqn{l_{i,\max,k} = \arg\max_{l=1,\dots,k} \{ med_{i,l} \}}

Two different numbers of clusters (\eqn{k_2 > k_1}) are then compared by 
their differences in CLR values for each hashtag between the top clusters,
\deqn{f_i = med_{i,l_{i,\max,k_2}} - med_{i,l_{i,\max,k_1}}, \quad k_2 > k_1}

If the maximum of \eqn{f_i} across all hashtags,
i.e., \deqn{\max_{i=1,\dots,N} \{ f_i \}},
exceeds a given threshold (default: 1), then \eqn{k_2} is recommended.
}
\examples{
\donttest{
hto.kmed.cl1 <- KmedCluster(hto.clr.mtx)
hto.kmed.cl2 <- KmedCluster(hto.clr.mtx, optional = TRUE, extra_cluster = 1)
ExamineClusterNumber(hto.clr.mtx, hto.kmed.cl1, hto.kmed.cl2)
}   

}
