% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMedoidsCluster.R
\name{KmedCluster}
\alias{KmedCluster}
\title{K-Medoids clustering}
\usage{
KmedCluster(
  clr.norm,
  fast_mode = FALSE,
  prop = 0.1,
  optional = FALSE,
  extra_cluster = 1
)
}
\arguments{
\item{clr.norm}{The local CLR normalized data comes from the output of 
\link{LocalCLRNorm} function.}

\item{fast_mode}{A logical value indicating whether to perform fast mode. If TRUE, 
CLARA clustering is performed; otherwise, PAM clustering is performed. 
Fast mode (TRUE) is recommended when there are more than 10,000 cells. 
Default: FALSE.}

\item{prop}{Proportion of cells used in CLARA clustering. This parameter will
only be used when fast_mode is TRUE. Default: 0.1.}

\item{optional}{A logical value indicating whether extra clusters should be created. 
If TRUE, optional clustering is performed, resulting in more clusters 
than the number of hashes. Default: FALSE.}

\item{extra_cluster}{The number of extra clusters. The total number of clusters 
after clustering will equal the number of hashtags plus the number of 
extra clusters. This parameter will only be used when the optional 
parameter is TRUE. Default: 1.}
}
\value{
A list of clustering information. 
\describe{
  \item{medoids}{Medoids of clusters.}
  \item{clustering}{Cluster of each cell.}
  \item{\dots}{Other outputs from the underlying `pam` or `clara` function, such as `id.med`, `objective`, `isolation`, etc.}
}
}
\description{
Perform K-Medoids clustering on the local CLR normalized data.
}
\details{
It is recommended to set `fast_mode` to TRUE when the hashing data contain more 
than 10,000 cells. In fast mode, CLARA clustering is performed, whereas in the 
default mode, PAM clustering is used. Since CLARA clustering uses only a 
proportion of cells, the `prop` parameter specifies the proportion of cells to 
be used for clustering.

Optional clustering is recommended when the number of clusters—equal to the 
number of hashtags—is not sufficient to achieve good demultiplexing results. 
This is particularly useful for low-quality data. If the number of clusters 
equal to the number of hashtags does not yield satisfactory demultiplexing, try 
gradually increasing the number of clusters by setting `extra_cluster` from 1 upward.
}
\examples{
\donttest{
hto.kmed.cl <- KmedCluster(hto.clr.mtx)
# Perform clustering in fast mode when there are more than 100,000 cells or when the default clustering is too slow
hto.kmed.cl <- KmedCluster(hto.clr.mtx, fast_mode = TRUE)
# PPerform optional clustering when more clusters than hashtags are required
hto.kmed.cl <- KmedCluster(hto.clr.mtx, optional = TRUE)
# Perform optional clustering by specifying the number of extra clusters
hto.kmed.cl <- KmedCluster(hto.clr.mtx, optional = TRUE, extra_cluster = 3)
hto.kmed.cl$medoids
head(hto.kmed.cl$clustering)
}

}
\references{
Maechler M, Rousseeuw P, Struyf A, Hubert M, Hornik K (2025). 
cluster: Cluster Analysis Basics and Extensions.
}
