% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{LabelClusterHTO}
\alias{LabelClusterHTO}
\title{Label each cluster to its HTO}
\usage{
LabelClusterHTO(clr.norm, kmed.cl, core.drop.assign, label_method = "medoids")
}
\arguments{
\item{clr.norm}{The local CLR normalized data comes from the output of the 
\link{LocalCLRNorm} function.}

\item{kmed.cl}{K-Medoids clustering results from the output of the 
\link{KmedCluster} function.}

\item{core.drop.assign}{Classification of core and non-core cells from the 
output of the \link{DefineNonCore} function.}

\item{label_method}{Specifies the method for labelling clusters. Options: 
"medoids" (assigns clusters based on cluster medoids) or "expression" 
(assigns clusters based on the maximum-expression hashtag). Default is 
"medoids".}
}
\value{
A named integer vector, where values represent clusters and names 
correspond to hashtags.
}
\description{
Labels each cluster according to its sample of origin.
}
\details{
In the medoids-based approach, a hashtag \eqn{i} is assigned to the cluster 
\eqn{l} whose medoid has the highest local CLR value \eqn{h}, i.e., 
\eqn{l = \arg\max_i \big(h_{i, M^{(l)}}\big)}.

In the expression-based approach, non-core cells are first excluded. A hashtag 
\eqn{i} is assigned to the cluster \eqn{l} with the highest median CLR value 
\eqn{h}, i.e., \eqn{l = \arg\max_i \big(\mathrm{med}_{i,l}\big)}.

When a cluster shows the maximum median local CLR value in two (or more) 
hashtags, this indicates conflicting hashtags, denoted \eqn{\mathcal{C}_i}. 
This means the cluster expresses multiple hashtags at high levels. To assign 
each conflicting hashtag \eqn{i} to a unique cluster, we calculate a normalized 
score \eqn{D}, defined as
\deqn{
D_i = \frac{\big| \mathrm{Med}(h_{i,l_1}) - \mathrm{Med}(h_{i,l_2}) \big|}
           { \tfrac{1}{2} \big( \mathrm{IQR}(h_{i,l_1}) + \mathrm{IQR}(h_{i,l_2}) \big)},
\quad i \in \mathcal{C}_i,
}
where \eqn{l_1} and \eqn{l_2} are the top two clusters with the highest 
median local CLR expression for hashtag \eqn{i}. The median local CLR value is 
normalized by the interquartile range (IQR) of each top cluster. The cluster 
is assigned to the hashtag with the maximum score \eqn{D}, i.e., 
\eqn{l = \arg\max_{i \in \mathcal{C}_i} D_i}. The remaining conflicting hashtags 
in \eqn{\mathcal{C}_i} are assigned to the unassigned clusters with the next 
highest median CLR expression.
}
\examples{
\donttest{
# Label clusters using medoids information
hto.cluster.assign <- LabelClusterHTO(hto.clr.mtx, hto.kmed.cl, hto.noncore, "medoids")
# Label clusters using expression information
hto.cluster.assign <- LabelClusterHTO(hto.clr.mtx, hto.kmed.cl, hto.noncore, "expression")
hto.cluster.assign
}

}
