% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{MDDensPlot}
\alias{MDDensPlot}
\title{Density plot of Mahalanobis distances}
\usage{
MDDensPlot(md.mat, md_cut_q = 0.95, md_cut = NULL, md_auto = FALSE)
}
\arguments{
\item{md.mat}{Matrix of Mahalanobis distance from the output of the 
\link{CalculateMD} function.}

\item{md_cut_q}{Numeric; cut-off to determine outlier cells (negatives and doublets)
based on the quantile of Mahalanobis distance. This parameter is indicated 
by the red line in the plot. Default is 0.95.}

\item{md_cut}{Numeric; cut-off to determine outlier cells (negatives and doublets)
based on the Mahalanobis distance directly. You only need to specify one of
\code{md_cut_q}, \code{md_cut}, or \code{md_auto} to determine the cut-offs.
This parameter is indicated by the red line in the plot. Default is NULL.}

\item{md_auto}{Logical; whether to use an automatic method to determine cut-offs
of outlier cells (negatives and doublets). If TRUE, the 0.975 quantile of 
a chi-square distribution will be used as the cut-off. You only need to
specify one of \code{md_cut_q}, \code{md_cut}, or \code{md_auto}. Default 
is FALSE.}
}
\value{
A density plot of Mahalanobis distances.
}
\description{
Draw a density plot of the minimum Mahalanobis distance distribution across 
all cells.
}
\examples{
\donttest{
# Visualize the cut-off for outlier cells based on the quantile of the Mahalanobis distance
MDDensPlot(hto.md.mat, md_cut_q = 0.94)
# Visualize the cut-off for outlier cells based on the value of Mahalanobis distance directly
MDDensPlot(hto.md.mat, md_cut = 5)
# Visualize the cut-off for outlier cells based on an automatic method
MDDensPlot(hto.md.mat, md_auto = TRUE)
}

}
