% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{OutlierHTOPlot}
\alias{OutlierHTOPlot}
\title{Density plot of hashtag library sizes of outlier cells}
\usage{
OutlierHTOPlot(hash.count, drop.outlier.assign, num_modes = 3, lib_cut = NULL)
}
\arguments{
\item{hash.count}{The raw hashing count matrix has rows representing hashtags 
and columns representing cells.}

\item{drop.outlier.assign}{A classification vector of outlier cells and singlets, 
from the output of the \link{AssignOutlierDrop} function.}

\item{num_modes}{Integer; the number of modes (including antimodes) in the 
distribution of hashtag library size, used to determine the cut-off 
between negatives and doublets. The modes and antimodes are displayed 
with red lines in the plot. Default is 3.}

\item{lib_cut}{Double; the cut-off value applied directly to hashtag library 
sizes. This parameter is indicated by a red line in the plot. It is an 
alternative to \code{num_modes} and should only be specified when 
\code{num_modes} does not provide an ideal cut-off. The default is NULL.}
}
\value{
A density plot of hashtag library sizes for outlier cells (negatives 
and doublets).
}
\description{
Draw a density plot of hashtag library sizes for outlier cells (negatives and 
doublets).
}
\examples{
\donttest{
OutlierHTOPlot(hto.mtx, hto.outlier.assign)
OutlierHTOPlot(hto.mtx, hto.outlier.assign, num_modes = 10)
# Specify the cut-off directly using hashtag library size
OutlierHTOPlot(hto.mtx, hto.outlier.assign, lib_cut= 5.6)
}

}
